Extract page settings to smaller components for readability
