+++ UI/Web/src/app/settings/settings/_components/directory-settings/directory-settings.component.html
@@ -1,0 +1,37 @@
+<div class="flex flex-row">
+  <app-form-input [formGroup]="pageForm" controlName="custom_root_dir" label="Custom Root Dir" class="w-full mr-5"></app-form-input>
+  <div (click)="updateCustomDir()" class="hover:cursor-pointer text-black dark:text-white flex flex-col justify-around">
+    <div></div>
+    <ng-icon name="heroSquare3Stack3d" size="32"></ng-icon>
+  </div>
+</div>
+
+<div class="flex flex-col space-y-2 mt-2">
+
+  <h1 class="w-2/5 border-b-2 border-black dark:border-white mb-2">Directories</h1>
+
+  @for (dir of getDirs(); track $index) {
+    <div class="flex flex-row justify-center items-center space-x-2">
+      <input
+        type="text"
+        (change)="updateDir($index, $event)"
+        class="focus:ring-primary-600 focus:border-primary-600 block w-full rounded-lg border border-gray-300 bg-gray-50 p-2.5 text-gray-900 sm:text-sm dark:border-gray-600 dark:bg-gray-700 dark:text-white dark:placeholder-gray-400 dark:focus:border-blue-500 dark:focus:ring-blue-500"
+        [value]="dir"
+      />
+      <div (click)="getNewDir($index)" class="hover:cursor-pointer text-black dark:text-white">
+        <ng-icon name="heroSquare3Stack3d" size="32"></ng-icon>
+      </div>
+
+      <div (click)="removeDir($index)" class="hover:cursor-pointer text-black dark:text-white">
+        <ng-icon name="heroMinus" size="32"></ng-icon>
+      </div>
+    </div>
+  }
+
+  <div class="w-full flex flex-row justify-end">
+    <div (click)="getNewDir(getDirs().length)" class="hover:cursor-pointer text-black dark:text-white">
+      <ng-icon name="heroPlus" size="32"></ng-icon>
+    </div>
+  </div>
+
+</div>
