+++ UI/Web/src/app/settings/settings/_components/modifier-settings/modifier-settings.component.html
@@ -1,0 +1,96 @@
+<div class="flex flex-col space-y-2 mt-2">
+
+  <div class="flex flex-row w-full justify-between">
+    <h1 class="w-2/5 border-b-2 border-black dark:border-white mb-2">Modifiers</h1>
+    <div class="flex flex-row space-x-2">
+      <div (click)="addModifier()" class="hover:cursor-pointer text-black dark:text-white">
+        <ng-icon name="heroPlus" size="32"></ng-icon>
+      </div>
+      <div (click)="toggleModifiers()" class="hover:cursor-pointer text-black dark:text-white">
+        <ng-icon [name]="showModifiers ? 'heroChevronDown' : 'heroChevronUp'" size="32"></ng-icon>
+      </div>
+    </div>
+  </div>
+
+  @if (showModifiers && !isMobile) {
+    <div class="flex flex-col md:grid md:grid-cols-2 space-y-2" [@dropDown]>
+
+      @for (m of getModifiers() | keyvalue; track m.key) {
+
+        <div class="flex flex-col space-y-2 md:m-2 p-2 bg-gray-100 dark:bg-gray-800 rounded-xl shadow dropdown-item">
+
+          <div class="flex flex-row w-full justify-between">
+            <h2 class="w-2/5 border-b-2 border-black dark:border-white mb-2">{{m.value.title}}</h2>
+            <div (click)="removeModifier(m.key)" class="hover:cursor-pointer text-black dark:text-white">
+              <ng-icon name="heroMinus" size="32"></ng-icon>
+            </div>
+          </div>
+
+          <div class="flex flex-row space-x-5 items-center">
+            <h3 class="text-left w-1/6">Title: </h3>
+            <input
+              type="text"
+              (change)="updateModifierTitle(m.key, $event)"
+              class="focus:ring-primary-600 focus:border-primary-600 overflow-x-auto block w-2/3 rounded-lg border border-gray-300 bg-gray-50 p-2.5 text-gray-900 sm:text-sm dark:border-gray-600 dark:bg-gray-700 dark:text-white dark:placeholder-gray-400 dark:focus:border-blue-500 dark:focus:ring-blue-500"
+              [value]="m.value.title"
+            />
+          </div>
+
+          <div class="flex flex-row space-x-5 items-center">
+            <h3 class="text-left w-1/6">Key: </h3>
+            <input
+              type="text"
+              (change)="updateModifierKey(m.key, $event)"
+              class="focus:ring-primary-600 focus:border-primary-600 overflow-x-auto block w-2/3 rounded-lg border border-gray-300 bg-gray-50 p-2.5 text-gray-900 sm:text-sm dark:border-gray-600 dark:bg-gray-700 dark:text-white dark:placeholder-gray-400 dark:focus:border-blue-500 dark:focus:ring-blue-500"
+              [value]="m.key"
+            />
+          </div>
+
+          <div class="max-h-48 flex flex-col space-y-2 overflow-y-auto mr-5">
+
+            <div class="flex flex-row">
+              <h3 class="text-left hidden md:block w-2/5">Keys </h3>
+              <h3 class="text-left hidden md:block w-2/5">Values </h3>
+            </div>
+
+            @for (v of m.value.values | keyvalue; track v.key) {
+
+              <div class="flex flex-row justify-between">
+                <div class="w-2/5">
+                  <h3 class="text-left block md:hidden">Key </h3>
+                  <input
+                    type="text"
+                    class="focus:ring-primary-600 focus:border-primary-600 overflow-x-auto block w-5/6 rounded-lg border border-gray-300 bg-gray-50 p-2.5 text-gray-900 sm:text-sm dark:border-gray-600 dark:bg-gray-700 dark:text-white dark:placeholder-gray-400 dark:focus:border-blue-500 dark:focus:ring-blue-500"
+                    [value]="v.key"
+                  />
+                </div>
+
+                <div class="w-2/5">
+                  <h3 class="text-left block md:hidden">Value </h3>
+                  <input
+                    type="text"
+                    class="focus:ring-primary-600 focus:border-primary-600 overflow-x-auto block w-5/6 rounded-lg border border-gray-300 bg-gray-50 p-2.5 text-gray-900 sm:text-sm dark:border-gray-600 dark:bg-gray-700 dark:text-white dark:placeholder-gray-400 dark:focus:border-blue-500 dark:focus:ring-blue-500"
+                    [value]="v.value"
+                  />
+                </div>
+
+                <div (click)="removeModifierValue(m.key, v.key)" class="hover:cursor-pointer text-black dark:text-white">
+                  <ng-icon name="heroMinus" size="32"></ng-icon>
+                </div>
+              </div>
+            }
+
+            <div class="flex flex-row justify-end">
+              <div (click)="addModifierValue(m.key)" class="hover:cursor-pointer text-black dark:text-white">
+                <ng-icon name="heroPlus" size="32"></ng-icon>
+              </div>
+            </div>
+
+          </div>
+        </div>
+      }
+    </div>
+  } @else if (showModifiers && isMobile) {
+    <div class="text-red-400 dark:text-red-800">Editing modifiers is not supported on mobile devices</div>
+  }
+</div>
